+++ src/qp_vault/embeddings/__init__.py
@@ -1,0 +1,8 @@
+# Copyright 2026 Quantum Pipes Technologies, LLC
+# SPDX-License-Identifier: Apache-2.0
+
+"""Built-in embedding providers for qp-vault."""
+
+from qp_vault.embeddings.noop import NoopEmbedder
+
+__all__ = ["NoopEmbedder"]